+++ populationsim/balancer/__init__.py
@@ -1,8 +1,10 @@
-from .simultaneous import SimultaneousListBalancer
-from .single import ListBalancer
-
+from .simul_balancer import SimultaneousListBalancer
+from .single_balancer import ListBalancer
+from .wrappers import do_balancing, do_simul_balancing
 
 __all__ = [
     "SimultaneousListBalancer",
     "ListBalancer",
+    "do_balancing",
+    "do_simul_balancing",
 ]
