+++ populationsim/balancer/single_balancer.py
@@ -5,11 +5,9 @@ import logging
 import numpy as np
 import pandas as pd
 
-from populationsim.core.config import setting
 from .balancers import np_balancer_py
 from .balancers_numba import np_balancer_numba
 from .constants import (
-    DEFAULT_MAX_ITERATIONS,
     MAX_INT,
     MIN_IMPORTANCE,
     MIN_CONTROL_VALUE,
@@ -185,118 +183,3 @@ class ListBalancer:
         ]
 
         return status, weights, controls
-
-
-def do_balancing(
-    control_spec,
-    total_hh_control_col,
-    max_expansion_factor,
-    min_expansion_factor,
-    absolute_upper_bound,
-    absolute_lower_bound,
-    incidence_df,
-    control_totals,
-    initial_weights,
-    use_hard_constraints,
-    use_numba,
-    numba_precision,
-):
-
-    # incidence table should only have control columns
-    incidence_df = incidence_df[control_spec.target]
-
-    # master_control_index is total_hh_control_col
-    if total_hh_control_col not in incidence_df.columns:
-        raise RuntimeError(
-            "total_hh_control column '%s' not found in incidence table"
-            % total_hh_control_col
-        )
-    total_hh_control_index = incidence_df.columns.get_loc(total_hh_control_col)
-
-    # control_totals series rows and incidence_df columns should be aligned
-    assert total_hh_control_index == control_totals.index.get_loc(total_hh_control_col)
-
-    control_totals = control_totals.values
-
-    control_importance_weights = control_spec.importance
-
-    if min_expansion_factor:
-
-        # number_of_households in this seed geograpy as specified in seed_controls
-        number_of_households = control_totals[total_hh_control_index]
-
-        total_weights = initial_weights.sum()
-        lb_ratio = (
-            min_expansion_factor * float(number_of_households) / float(total_weights)
-        )
-
-        # Added hard limit of min_expansion_factor value that would otherwise drift
-        # due to the float(number_of_households) / float(total_weights) calculation
-        if use_hard_constraints:
-            lb_ratio = np.clip(lb_ratio, a_min=min_expansion_factor, a_max=None)
-
-        lb_weights = initial_weights * lb_ratio
-
-        if absolute_lower_bound:
-            lb_weights = lb_weights.clip(lower=absolute_lower_bound)
-        else:
-            lb_weights = lb_weights.clip(lower=0)
-
-    elif absolute_lower_bound:
-        lb_weights = initial_weights.clip(lower=absolute_lower_bound)
-
-    else:
-        lb_weights = None
-
-    if max_expansion_factor:
-
-        # number_of_households in this seed geograpy as specified in seed_controlss
-        number_of_households = control_totals[total_hh_control_index]
-
-        total_weights = initial_weights.sum()
-        ub_ratio = (
-            max_expansion_factor * float(number_of_households) / float(total_weights)
-        )
-
-        # Added hard limit of max_expansion_factor value that would otherwise drift
-        # due to the float(number_of_households) / float(total_weights) calculation
-        if use_hard_constraints:
-            ub_ratio = np.clip(ub_ratio, a_max=max_expansion_factor, a_min=None)
-
-        ub_weights = initial_weights * ub_ratio
-
-        if absolute_upper_bound:
-            ub_weights = (
-                ub_weights.round().clip(upper=absolute_upper_bound, lower=1).astype(int)
-            )
-        else:
-            ub_weights = ub_weights.round().clip(lower=1).astype(int)
-
-    elif absolute_upper_bound:
-        ub_weights = (
-            ub_weights.round().clip(upper=absolute_upper_bound, lower=1).astype(int)
-        )
-
-    else:
-        ub_weights = None
-
-    max_iterations = setting(
-        "MAX_BALANCE_ITERATIONS_SEQUENTIAL", DEFAULT_MAX_ITERATIONS
-    )
-
-    balancer = ListBalancer(
-        incidence_table=incidence_df,
-        initial_weights=initial_weights,
-        control_totals=control_totals,
-        control_importance_weights=control_importance_weights,
-        lb_weights=lb_weights,
-        ub_weights=ub_weights,
-        master_control_index=total_hh_control_index,
-        max_iterations=max_iterations,
-        use_numba=use_numba,
-        numba_precision=numba_precision,
-    )
-
-    status, weights, controls = balancer.balance()
-
-    return status, weights, controls
