+++ populationsim/integerizer/__init__.py
@@ -1,3 +1,7 @@
-from .integerizer import do_integerizing
+from .wrappers import do_integerizing, do_simul_integerizing, do_sequential_integerizing
 
-__all__ = ["do_integerizing"]
+__all__ = [
+    "do_integerizing",
+    "do_simul_integerizing",
+    "do_sequential_integerizing",
+]
