+++ populationsim/integerizer/integerizer.py
@@ -5,53 +5,12 @@ import logging
 
 import numpy as np
 import pandas as pd
-from populationsim.core.config import setting
 
 from .lp import get_single_integerizer
-from .lp import STATUS_SUCCESS, STATUS_OPTIMAL
-
+from .lp import STATUS_OPTIMAL
+from .smart_round import smart_round
 
 logger = logging.getLogger(__name__)
-
-
-def smart_round(int_weights, resid_weights, target_sum):
-    """
-    Round weights while ensuring (as far as possible that result sums to target_sum)
-
-    Parameters
-    ----------
-    int_weights : numpy.ndarray(int)
-    resid_weights : numpy.ndarray(float)
-    target_sum : int
-
-    Returns
-    -------
-    rounded_weights : numpy.ndarray array of ints
-    """
-    assert len(int_weights) == len(resid_weights)
-    assert (int_weights == int_weights.astype(int)).all()
-    assert target_sum == int(target_sum)
-
-    target_sum = int(target_sum)
-
-    # integer part of numbers to round (astype both copies and coerces)
-    rounded_weights = int_weights.astype(int)
-
-    # find number of residuals that we need to round up
-    int_shortfall = target_sum - rounded_weights.sum()
-
-    # clip to feasible, in case target was not achievable by rounding
-    int_shortfall = np.clip(int_shortfall, 0, len(resid_weights))
-
-    # Order the residual weights and round at the tipping point where target_sum is achieved
-    if int_shortfall > 0:
-        # indices of the int_shortfall highest resid_weights
-        i = np.argsort(resid_weights)[-int_shortfall:]
-
-        # add 1 to the integer weights that we want to round upwards
-        rounded_weights[i] += 1
-
-    return rounded_weights
 
 
 class Integerizer:
@@ -211,150 +170,3 @@ class Integerizer:
         )
 
         return status
-
-
-def do_integerizing(
-    trace_label,
-    control_spec,
-    control_totals,
-    incidence_table,
-    float_weights,
-    total_hh_control_col,
-):
-    """
-
-    Parameters
-    ----------
-    trace_label : str
-        trace label indicating geography zone being integerized (e.g. PUMA_600)
-    control_spec : pandas.Dataframe
-        full control spec with columns 'target', 'seed_table', 'importance', ...
-    control_totals : pandas.Series
-        control totals explicitly specified for this zone
-    incidence_table : pandas.Dataframe
-    float_weights : pandas.Series
-        balanced float weights to integerize
-    total_hh_control_col : str
-        name of total_hh column (preferentially constrain to match this control)
-
-    Returns
-    -------
-    integerized_weights : pandas.Series
-    status : str
-        as defined in integerizer.STATUS_TEXT and STATUS_SUCCESS
-    """
-
-    # incidence table should only have control columns
-    incidence_table = incidence_table[control_spec.target]
-
-    if total_hh_control_col not in incidence_table.columns:
-        raise RuntimeError(
-            "total_hh_control column '%s' not found in incidence table"
-            % total_hh_control_col
-        )
-
-    zero_weight_rows = float_weights == 0
-    if zero_weight_rows.any():
-        logger.debug(
-            "omitting %s zero weight rows out of %s"
-            % (zero_weight_rows.sum(), len(incidence_table.index))
-        )
-        incidence_table = incidence_table[~zero_weight_rows]
-        float_weights = float_weights[~zero_weight_rows]
-
-    total_hh_control_value = control_totals[total_hh_control_col]
-
-    status = None
-    if setting("INTEGERIZE_WITH_BACKSTOPPED_CONTROLS") and len(control_totals) < len(
-        incidence_table.columns
-    ):
-
-        ##########################################
-        # - backstopped control_totals
-        # Use balanced float weights to establish target values for all control values
-        # note: this more frequently results in infeasible solver results
-        ##########################################
-
-        relaxed_control_totals = np.round(
-            np.dot(np.asanyarray(float_weights), incidence_table.values)
-        )
-        relaxed_control_totals = pd.Series(
-            relaxed_control_totals, index=incidence_table.columns.values
-        )
-
-        # if the incidence table has only one record, then the final integer weights
-        # should be just an array with 1 element equal to the total number of households;
-        assert len(incidence_table.index) > 1
-
-        integerizer = Integerizer(
-            incidence_table=incidence_table,
-            control_importance_weights=control_spec.importance,
-            float_weights=float_weights,
-            relaxed_control_totals=relaxed_control_totals,
-            total_hh_control_value=total_hh_control_value,
-            total_hh_control_index=incidence_table.columns.get_loc(
-                total_hh_control_col
-            ),
-            control_is_hh_based=control_spec["seed_table"] == "households",
-            trace_label="backstopped_%s" % trace_label,
-        )
-
-        # otherwise, solve for the integer weights using the Mixed Integer Programming solver.
-        status = integerizer.integerize()
-
-        logger.debug(
-            "Integerizer status for backstopped %s: %s" % (trace_label, status)
-        )
-
-    # if we either tried backstopped controls or failed, or never tried at all
-    if status not in STATUS_SUCCESS:
-
-        ##########################################
-        # - unbackstopped partial control_totals
-        # Use balanced weights to establish control totals only for explicitly specified controls
-        # note: this usually results in feasible solver results, except for some single hh zones
-        ##########################################
-
-        balanced_control_cols = control_totals.index
-        incidence_table = incidence_table[balanced_control_cols]
-        control_spec = control_spec[control_spec.target.isin(balanced_control_cols)]
-
-        relaxed_control_totals = np.round(
-            np.dot(np.asanyarray(float_weights), incidence_table.values)
-        )
-        relaxed_control_totals = pd.Series(
-            relaxed_control_totals, index=incidence_table.columns.values
-        )
-
-        integerizer = Integerizer(
-            incidence_table=incidence_table,
-            control_importance_weights=control_spec.importance,
-            float_weights=float_weights,
-            relaxed_control_totals=relaxed_control_totals,
-            total_hh_control_value=total_hh_control_value,
-            total_hh_control_index=incidence_table.columns.get_loc(
-                total_hh_control_col
-            ),
-            control_is_hh_based=control_spec["seed_table"] == "households",
-            trace_label=trace_label,
-        )
-
-        status = integerizer.integerize()
-
-        logger.debug(
-            "Integerizer status for unbackstopped %s: %s" % (trace_label, status)
-        )
-
-    if status not in STATUS_SUCCESS:
-        logger.error(
-            "Integerizer failed for %s status %s. "
-            "Returning smart-rounded original weights" % (trace_label, status)
-        )
-    elif status != "OPTIMAL":
-        logger.warning(
-            "Integerizer status non-optimal for %s status %s." % (trace_label, status)
-        )
-
-    integerized_weights = pd.Series(0, index=zero_weight_rows.index)
-    integerized_weights.update(integerizer.weights["integerized_weight"])
-    return integerized_weights, status
