+++ populationsim/steps/final_seed_balancing.py
@@ -5,9 +5,12 @@ import logging
 import pandas as pd
 
 from populationsim.core import inject, config
-
-from populationsim.helper import get_weight_table, get_control_table, weight_table_name
-from populationsim.balancer.single import do_balancing
+from populationsim.core.helper import (
+    get_weight_table,
+    get_control_table,
+    weight_table_name,
+)
+from populationsim.balancer import do_balancing
 
 
 logger = logging.getLogger(__name__)
