+++ populationsim/steps/repop_balancing.py
@@ -5,9 +5,12 @@ import logging
 import pandas as pd
 
 from populationsim.core import inject, config
-
-from populationsim.helper import get_control_table, weight_table_name, get_weight_table
-from populationsim.balancer.single import do_balancing
+from populationsim.core.helper import (
+    get_control_table,
+    weight_table_name,
+    get_weight_table,
+)
+from populationsim.balancer import do_balancing
 from populationsim.integerizer import do_integerizing
